+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - 2. Paper
   - 3. Scissors
   - 4. Spock Hand
+  - 5. OK hand
 
 ---