+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - 3. Scissors
   - 4. Spock Hand
   - 5. OK hand
+- Adding Yaggababble
 
 ---